feat(condo): DOMA-10978 fix build by fix including of constants
