+++ apps/condo/domains/billing/gql.js
@@ -8,7 +8,7 @@ const { gql } = require('graphql-tag')
 
 const { generateGqlQueries } = require('@open-condo/codegen/generate.gql')
 
-const { RECIPIENT_FIELDS_DEFINITION } = require('@condo/domains/acquiring/schema/fields/Recipient')
+const { RECIPIENT_FIELDS_DEFINITION } = require('@condo/domains/acquiring/constants/gql')
 const { ADDRESS_META_SUBFIELDS_QUERY_LIST } = require('@condo/domains/property/schema/fields/AddressMetaField')
 
 const COMMON_FIELDS = 'id dv sender { dv fingerprint } v deletedAt newId createdBy { id name } updatedBy { id name } createdAt updatedAt'
